+++ src/utils/__tests__/labels.spec.js
@@ -1,0 +1,15 @@
+import { labelLayer } from '../labels'
+import defaults from '../style'
+
+const id = 'abc'
+const opacity = 0.5
+
+describe('labels', () => {
+    it('Should set opacity for for label layer', () => {
+        expect(labelLayer({ id, opacity }).paint['text-opacity']).toBe(opacity)
+    })
+
+    it('Should set default opacity for label layer', () => {
+        expect(labelLayer({ id }).paint['text-opacity']).toBe(defaults.opacity)
+    })
+})